+++ lib/src/util/pageable/pageable.dart
@@ -1,4 +1,5 @@
 export 'inconsistent_page_data_exception.dart';
 export 'offset_pageable_resource.dart';
 export 'pageable_data.dart';
+export 'size_pageable_data.dart';
 export 'size_pageable_resource.dart';